Build bustage fix, part 3.
UltraBlame original commit: dafe0f926f50ea44b00848ca93837c44e3d12ae8

+++ js/src/ion/shared/CodeGenerator-x86-shared.cpp
@@ -62,7 +62,7 @@ class DeferredJumpTable : public DeferredData
         void **jumpData = (void **)buffer;
 
         
-        for (uint j = 0; j < lswitch->mir()->numCases(); j++) { 
+        for (size_t j = 0; j < lswitch->mir()->numCases(); j++) { 
             LBlock *caseblock = lswitch->mir()->getCase(j)->lir();
             Label *caseheader = caseblock->label();
